requirements.txt
openai
langchain
python-dotenv
streamlit
faiss-cpu
pypdf